microservices/LLM-Service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.8.1
openai==1.3.0
aiohttp==3.9.0
python-multipart==0.0.6
python-dotenv==1.0.0
sqlalchemy==2.0.23
asyncpg==0.29.0
redis==5.0.1
numpy==1.25.2
pandas==2.1.3
scikit-learn==1.3.2
PyJWT==2.8.0
cryptography==41.0.7
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
